requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
transformers
python_dotenv